+++ Drivers/WS2812/Src/color_values.c
@@ -23,7 +23,7 @@
 // Look up table for led color bit patterns.  "Waste" 4k of flash but is a
 // lot faster (not measured accurately but I'd say about double) than bit
 // manipulation.  I'd love to hear if someone got a better idea ;)
-const uint16_t color_value[256][8] = {
+const uint32_t color_value[256][8] = {
         { LED_OFF, LED_OFF, LED_OFF, LED_OFF, LED_OFF, LED_OFF, LED_OFF, LED_OFF },
         { LED_OFF, LED_OFF, LED_OFF, LED_OFF, LED_OFF, LED_OFF, LED_OFF, LED_ON },
         { LED_OFF, LED_OFF, LED_OFF, LED_OFF, LED_OFF, LED_OFF, LED_ON, LED_OFF },
